+++ src/main/java/duke/Duke.java
@@ -1,0 +1,54 @@
+package duke;
+
+import duke.command.Command;
+import duke.exception.DukeException;
+import duke.util.Parser;
+import duke.util.Storage;
+import duke.util.TaskList;
+import duke.util.Ui;
+
+public class Duke {
+
+    private Storage storage;
+    private TaskList tasks;
+    private Ui ui;
+
+    public Duke(String file, String fileParent, String name, String logo) {
+        ui = new Ui(name, logo, System.in);
+        storage = new Storage(file, fileParent);
+        try {
+            tasks = new TaskList(storage.readFromFile());
+        } catch (DukeException e) {
+            ui.showLoadingError();
+            tasks = new TaskList();
+        }
+    }
+
+    public void run() {
+        ui.showWelcome();
+        boolean isExit = false;
+        while (!isExit) {
+            try {
+                String fullCommand = ui.readCommand();
+                ui.showLine();
+                Command c = Parser.parseCommand(fullCommand);
+                c.execute(tasks, ui, storage);
+                isExit = c.isExit();
+            } catch (DukeException e) {
+                ui.showError(e.getMessage());
+            } finally {
+                ui.showLine();
+            }
+        }
+    }
+
+
+    public static void main(String[] args) {
+        String logo = "\t  ____   __  _  _  __    ___  ___  ____  ___  ____\n\t"
+                + " (_  _) /__\\( \\/ )/__\\  / __)/ __)(_  _)/ __)(_  _)\n\t"
+                + " .-_)( /(__)\\\\  //(__)\\ \\__ \\\\__ \\ _)(_ \\__ \\  )(\n\t"
+                + "\\____)(__)(__)\\/(__)(__)(___/(___/(____)(___/ (__)\n";
+
+        new Duke("duke.Duke.txt", "./data", "JavAssist", logo).run();
+    }
+}
