+++ src/main/java/duke/command/DeadlineCommand.java
@@ -1,26 +1,32 @@
+package duke.command;
+
+import duke.exception.DukeException;
+import duke.task.Deadline;
+import duke.task.Task;
+import duke.util.Storage;
+import duke.util.TaskList;
+import duke.util.Ui;
+
 import java.time.format.DateTimeParseException;
 
-public class EventCommand implements Command {
-
+public class DeadlineCommand implements Command {
     private String input;
 
-    public EventCommand(String input) {
+    public DeadlineCommand(String input) {
         this.input = input;
     }
 
     @Override
     public void execute(TaskList list, Ui ui, Storage storage) throws DukeException {
-        String[] task = input.split("/from");
-        String[] time = task[1].split("/to");
+        String[] task = input.split("/by");
         try {
-            Task t = new Event(task[0].substring(6).trim(), time[0].trim(), time[1].trim());
+            Task t = new Deadline(task[0].substring(9).trim(), task[1].trim());
             list.add(t);
             ui.showAdded(t, list);
             storage.writeToFile(list);
         } catch (DateTimeParseException e) {
             throw new DukeException(e.getMessage());
         }
-
     }
 
     @Override
