+++ src/main/java/duke/command/UnknownCommand.java
@@ -1,3 +1,10 @@
+package duke.command;
+
+import duke.exception.DukeException;
+import duke.util.Storage;
+import duke.util.TaskList;
+import duke.util.Ui;
+
 public class UnknownCommand implements Command {
     @Override
     public void execute(TaskList list, Ui ui, Storage storage) throws DukeException {
