+++ src/main/java/duke/task/Deadline.java
@@ -1,8 +1,10 @@
+package duke.task;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
 
-public class Deadline extends Task{
+public class Deadline extends Task {
     private LocalDateTime date;
 
     public Deadline(String task, String date) {
