+++ src/main/java/duke/task/Event.java
@@ -1,8 +1,10 @@
+package duke.task;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
 
-public class Event extends Task{
+public class Event extends Task {
     private LocalDateTime start;
     private LocalDateTime end;
