+++ src/main/java/duke/task/Todo.java
@@ -1,4 +1,6 @@
-public class Todo extends Task{
+package duke.task;
+
+public class Todo extends Task {
     public Todo(String task) {
         super(task);
     }
