+++ src/main/java/duke/util/Parser.java
@@ -1,3 +1,16 @@
+package duke.util;
+
+import duke.command.Command;
+import duke.command.DeadlineCommand;
+import duke.command.DeleteCommand;
+import duke.command.ExitCommand;
+import duke.command.EventCommand;
+import duke.command.ListCommand;
+import duke.command.MarkCommand;
+import duke.command.TodoCommand;
+import duke.command.UnknownCommand;
+import duke.exception.DukeException;
+
 import java.util.regex.Pattern;
 
 public class Parser {
